Move HAL serializer functionality into mixin
